SophiaProject/requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.4
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.1
requests==2.32.4
ruff==0.12.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
